tests.nixpkgs-check-by-name: Separate file for all problems
And introduce a function for some smaller indentation
